move NavLayout styles to utils/styles
