+++ components/NavLayout/NavLayout.js
@@ -1,6 +1,6 @@
 import SideBar from './SideBar';
 import TopBar from './TopBar';
-import styles from './NavLayoutStyles/NavLayout.module.css';
+import styles from '../../utils/styles/NavLayoutStyles/NavLayout.module.css';
 
 import { useAuthContext } from '../../utils/context/AuthProvider';
